requirements.txt
boto3~=1.34.109
python-dotenv~=1.0.1